refactor: :recycle: group path functions into global vs local
