+++ tests/core/test_path_local.py
@@ -4,6 +4,8 @@ from re import escape
 from pytest import fixture, mark, raises
 
 from seedcase_sprout.core import (
+    path_properties,
+    path_readme,
     path_resource,
     path_resource_data,
     path_resource_raw,
@@ -14,6 +16,33 @@ from tests.core.directory_structure_setup import (
     create_test_package_structure,
     create_test_resource_structure,
 )
+
+
+@mark.parametrize(
+    "function, expected_path",
+    [
+        (path_properties, "packages/1/datapackage.json"),
+        (path_readme, "packages/1/README.md"),
+    ],
+)
+def test_path_package_functions_return_expected_path(
+    tmp_sprout_global, function, expected_path
+):
+    # When, then
+    assert function(package_id=1) == tmp_sprout_global / expected_path
+    assert function(package_id=1).is_file() or function(package_id=1).is_dir()
+
+
+@mark.parametrize(
+    "function",
+    [path_properties, path_readme],
+)
+def test_path_package_functions_raise_error_if_package_id_does_not_exist(
+    tmp_sprout_global, function
+):
+    # When, then
+    with raises(NotADirectoryError, match=escape("[1]")):
+        function(package_id=2)
 
 
 # Given a package with two resources
